requirements.txt
fastapi~=0.104.1
pydantic==1.10.10
starlette~=0.27.0
python-dotenv~=1.0.0
openai==1.2.3
SQLAlchemy~=2.0.23
psycopg2-binary==2.9.5
pandas~=2.1.2